# [libc++] Validate vector<bool> copy/move-assignment operators in realistic scenarios