The existing tests for `vector<bool>` copy- and move-assignment operators are limited to 3 bits only, which are inadequate to cover realistic scenarios. Most `vector<bool>` operations have code paths that are executed only when multiple storage words are involved, with each storage word typically comprising 64 bits on a 64-bit platform. Furthermore, the existing tests fail to cover all combinations `POCCA`/`POCMA`, along with different allocator equality and/or reallocation scenarios, leaving some critical code paths untested. 

This patch enhances the test coverage by introducing new tests covering up to 5 storage words, ensuring that partial words in the front or tail, and whole words in the middle are all properly tested. Moreover, these new tests ensure that the copy- and move-assignment operators are tested under all combinations of `POCCA`/`POCMA` and various allocator equality scenarios, both with or without reallocations. 
 
(Note: The performance improvements originally intended have been dropped from this patch, as they have been addressed in separate work. This patch now focuses exclusively on improving the test coverage.)


~~### General description~~
~~This PR is part of a series aimed at improving the performance of `vector<bool>`. Each PR focuses on enhancing a specific subset of operations, ensuring they are self-contained and easy to review. The main idea for performance improvements involves using word-wise implementation along with bit manipulation techniques, rather than solely using bit-wise operations in the previous implementation, resulting in substantial performance gains.~~



~~### Current PR~~
~~This PR significantly enhances the performance of the `vector<bool>` move-assignment operator by at least **500x** in the specific scenario where `pocma = false_type` and the allocators compare unequal, where the move-assignment effectively performs element-wise copy-assignment.~~

~~While the previous implementation copied individual bits in a bit-wise manner, the new implementation copies entire storage words, yielding substantial performance improvements. Since this word-wise copying is essentially the same operation currently performed in the copy-assignment operator, we have refactored the common copy-operation into a function called `__copy_by_words`, which is now called by both the copy- and move-assignment operators.~~

~~### Tests~~

~~The existing tests for `vector<bool>` are largely insufficient to cover realistic scenarios. Existing tests are limited to cases of up to 1~2 bytes. However, most `vector<bool>` operations have code paths that are only executed when multiple storage words, 8 bytes each, are involved. Furthermore, the tests for copy- and move-assignment operators do not cover different combinations of `pocca`/`pocma` values with reallocation scenarios, leaving some important code paths untested.~~

~~To improve the test coverage, this PR completely rewrites the tests for  the copy/move- assignment operators, ensuring that every possible code path is properly tested.~~

~~Furthermore, this PR conducts benchmark testing for the move-assignment operator, demonstrating at least 500x  performance improvement under different word sizes of 32 and 64.~~


~~#### Before:~~

~~------------------------------------------------------------------------------------------~~
~~Benchmark                                                Time             CPU   Iterations~~
~~------------------------------------------------------------------------------------------~~
~~BM_Move_Assignment/vector_bool_uint32_t/5140480    7530901 ns      7545633 ns           86~~
~~BM_Move_Assignment/vector_bool_uint64_t/5140480    7677517 ns      7696633 ns           89~~
~~BM_Move_Assignment/vector_bool_size_t/5140480      7806975 ns      7828858 ns           85~~

~~####  After:~~

~~------------------------------------------------------------------------------------------~~
~~Benchmark                                                Time             CPU   Iterations~~
~~------------------------------------------------------------------------------------------~~
~~BM_Move_Assignment/vector_bool_uint32_t/5140480      14651 ns        14496 ns        45996~~
~~BM_Move_Assignment/vector_bool_uint64_t/5140480      14516 ns        14389 ns        47239~~
~~BM_Move_Assignment/vector_bool_size_t/5140480        15150 ns        15045 ns        44761~~